ns: move IApplicationVersionInterface
